Families update in order of creation
Before this commit, IFamilies (created from Nodes) will have their
onAdded/onRemoved sequence handled by a dictionary loop.

This does not preserve the order in which families are created and can
(and have) caused issues.

This commit fixes that so that the order of add/remove dispatching
happens in the order of family creation.